User Account

+++ lib/app/data/model/response_book_get.dart
@@ -49,7 +49,7 @@ class DataBook {
     publisher = json['publisher'];
     synopsis = json['synopsis'];
     rented = json['rented'];
-    rating = json['rating'];
+    rating = json['rating'].toDouble();
     status = json['status'];
     publishYear = json['publish_year'];
     createdAt = json['created_at'];
